ModuleMaster, TemplateTypeMaster Files
ModuleMaster, TemplateTypeMaster Files

+++ SchoolAPI/Models/TblTemplateTypeMaster.cs
@@ -23,6 +23,6 @@ namespace SchoolAPI.Models
 
         public DateTime? ModifiedDate { get; set; }
 
-        public long? Status { get; set; }
+        public int? Status { get; set; }
     }
 }
